+++ src/main/java/seedu/address/model/common/Mark.java
@@ -37,8 +37,8 @@ public class Mark implements Comparable<Mark> {
         return this.getValue().compareTo(other.getValue());
     }
 
-    public Float getValue() {
-        return Float.valueOf(this.internalString);
+    public Double getValue() {
+        return Double.valueOf(this.internalString);
     }
 
     @Override
