+++ src/main/java/seedu/address/ui/MoreDetailsPanel.java
@@ -145,11 +145,11 @@ public class MoreDetailsPanel extends UiPart<Region> {
         label.setStyle(style);
         components.add(label, 3, 0);
 
-        float assignmentWeight;
-        float assignmentMark;
-        float assignmentMaxMark;
-        float totalWeight = 0;
-        float weightedMarks = 0;
+        double assignmentWeight;
+        double assignmentMark;
+        double assignmentMaxMark;
+        double totalWeight = 0;
+        double weightedMarks = 0;
 
         int row = 1;
         Assignment assignment;
